separate components in transcripts route
