+++ arc3-api/src/auth/middlewares/whitelist.js
@@ -1,6 +1,7 @@
 
 const moderators = [
   '925408241951846450',
+  '794788752038952970',
   '1114651805155082371',
   '397233243856044037',
   '364180076297846790',
